Rename TextDifferencer classes.
